requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
bcrypt==4.0.1
cffi==1.17.1
click-didyoumean==0.3.1
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
jose==1.0.0
Mako==1.3.8
MarkupSafe==3.0.2
mypy-extensions==1.0.0
p==1.5.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
python-jose==3.3.0
python-multipart==0.0.19
python-slugify==8.0.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
text-unidecode==1.3
typing_extensions==4.12.2
uvicorn==0.32.1
